+++ client/src/Components/FormEvent/FormEvent.jsx
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
 import { useSelector } from 'react-redux'
-import validation from "../../utils/ValidationEventForm/validation";
+import validation from "../../Utils/ValidationEventForm/validation";
 import { useDispatch } from "react-redux";
-import { createEvent } from "../../redux/Action/action";
+import { createEvent } from "../../Redux/Action/action";
 import style from './formEvent.module.css'
 
 export default function FormEvent() {
