+++ client/src/Redux/Action/action-type.js
@@ -1,2 +1,3 @@
-export const GET_ALL_EVENT = 'GET_ALL_EVENT'
-export const CREATE_EVENT = 'CREATE_EVENT'
+export const GET_ALL_EVENT = 'GET_ALL_EVENT';
+export const CREATE_EVENT = 'CREATE_EVENT';
+export const GET_DETAIL = 'GET_DETAIL';
