--- client/src/Redux/store.js
@@ -1,10 +1,0 @@
-import { applyMiddleware, createStore } from "redux";
-import { composeWithDevTools } from 'redux-devtools-extension';
-import rootReducer from './Reducer/index'
-import thunk from 'redux-thunk';
-
-const store = createStore(
-    rootReducer, 
-    composeWithDevTools(applyMiddleware(thunk)))
-
-export default store
