+++ client/src/components/register/register.jsx
@@ -4,7 +4,7 @@ import registerValidation from "../../functions/Validations/registerValidation/v
 import axios from "axios";
 import styles from './register.module.css'
 import { Link } from "react-router-dom";
-import video from "../../utils/videos/backgroundLogin.mp4"
+import video from "../../Utils/videos/backgroundLogin.mp4"
 
 export default function Register() {
     const navigate = useNavigate()
