+++ client/src/index.js
@@ -6,7 +6,7 @@ import reportWebVitals from './reportWebVitals';
 
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import store from './redux/store/store';
+import store from "./Redux/store/store";
 import axios from 'axios';
 
 //axios.defaults.baseURL = "https://localhost:3001"
