--- client/src/redux/actions_type/actions_type.js
@@ -1,1 +1,0 @@
-export const GETDETAIL = "GETDETAIL"
