+++ client/src/redux/reducer/reducer.js
@@ -1,30 +1,22 @@
-import { GETDETAIL } from "../actions_type/actions_type";
+import { GET_ALL_EVENT, CREATE_EVENT, GET_DETAIL } from "../Action/action-type";
 
 const initialState = {
-  userSesion: {
-    //userId,
-    //userType
-  },
-  Detail: {},
-};
+    country: ['Argentina', 'Colombia', 'Venezuela', 'Uruguay'],
+    eventTypes: ['Musical', 'Deportivo', 'Artistico', 'Otro'],
+    events: [],
+    Detail: {},
+}
 
-const rootReducer = (state = initialState, action) => {
-  switch (action.type) {
-    case "USERSESION":
-      return {
-        ...state,
-        userSesion: action.payload,
-      };
-    case GETDETAIL:
-        console.log(state,action)
-      return {
-        ...state,
-        Detail: {...state.Detail,...action.payload,},
-      };
-
-    default:
-      return { ...state };
-  }
-};
-
-export default rootReducer;
+const rootReducer = (state = initialState, action)=>{
+    const {type, payload} = action
+    switch(type){
+        case GET_ALL_EVENT:
+            return {...state, events: payload};
+        case CREATE_EVENT:
+            return {...state, events: payload};
+        case GET_DETAIL:
+            return {...state, Detail: {...state.Detail,...action.payload}};
+        default: return state;
+    }
+}
+export default rootReducer
