+++ client/src/redux/store/store.js
@@ -1,7 +1,7 @@
 import { applyMiddleware, legacy_createStore } from 'redux';
 import thunk from 'redux-thunk';
 import { composeWithDevTools } from 'redux-devtools-extension'
-import rootReducer from '../reducer/reducer';
+import rootReducer from "../Reducer/reducer";
 
 const composedEnhancer = composeWithDevTools(applyMiddleware(thunk))
